add apiKey and appId to provider
