+++ package.json
@@ -50,7 +50,6 @@
     "@types/react": "17.0.34",
     "@types/react-modal": "^3.13.1",
     "@types/react-redux": "^7.1.23",
-    "@types/react-router-dom": "^5.3.2",
     "babel-jest": "27.4.5",
     "babel-loader": "8.2.3",
     "identity-obj-proxy": "3.0.0",
@@ -58,7 +57,6 @@
     "postcss": "8.4.5",
     "react": "17.0.2",
     "react-dom": "17.0.2",
-    "react-router-dom": "6.3.0",
     "regenerator-runtime": "0.13.9",
     "rollup": "2.60.0",
     "rollup-plugin-dts": "4.0.1",
@@ -73,7 +71,6 @@
     "react": ">=17.0.2",
     "react-dom": ">=17.0.2",
     "react-redux": "^>=17.2.8",
-    "react-router-dom": "^>=6.3.0",
     "@walletconnect/web3-provider": "1.7.1"
   },
   "scripts": {
