+++ src/components/Form/Form.tsx
@@ -3,8 +3,7 @@ import Modal from 'react-modal';
 
 import { CHAINID_NETWORK_MAP } from "../../services/service-constants";
 import { Web3Service } from "../../services/Web3Service/Web3Service";
-import { FormSignatureData, ACTION_TPYE, ErrorMessageData, Providers, VerifactionType } from "../../interface/web3-data-interface";
-import { FormProps } from '../../interface/form-interface';
+import { FormSignatureData, ACTION_TPYE, ErrorMessageData, Providers, VerifactionType, Backend } from "../../interface/web3-data-interface";
 
 import metamaskLogo from '../../assets/metamask.png';
 import walletConnectLogo from '../../assets/walletConnect.png';
@@ -17,6 +16,21 @@ declare global {
     ethereum: any
     web3: any;
   }
+}
+
+export interface FormProps {
+  primary: boolean;
+  backgroundcolor: string; 
+  size: string; 
+  formDataCallback: (web3Data: FormSignatureData) => void;
+  formErrorCallback: (errorData: ErrorMessageData) => void;
+  dappname: string;
+  logourl: string;
+  infuraId: string;
+  homePageurl: string;
+  disableErrorDisplay: boolean;
+  messageToSign: string;
+  backend?: Backend
 }
 
 /**
