--- src/hooks/useWeb3Data/index.ts
@@ -1,1 +1,0 @@
-export * from "./userWeb3Data";
