+++ src/index.ts
@@ -1,5 +1,4 @@
 // Export all component that you want any application using this library to be able to import
 export * from './components/Form';
 export * from './components/Button';
-export * from './hooks/AuthProvider';
-export * from './app';
+export * from './context/SecureAuth3Provider';
