--- src/interface/network-interface.ts
@@ -1,4 +1,0 @@
-export interface NetworkInfo {
-    name: string;
-    scannerUrl: string;
-}
