--- src/utils/routes.contant.ts
@@ -1,1 +1,0 @@
-export const API_PREFIX = process.env.NODE_ENV === 'production' ?  'https://api.secureauth3.com' : 'http://localhost:8080';
